requirements.txt
playwright
pytest
pytest-playwright
pyyaml
python-dotenv
